LayoutTests/fast/dom/DeviceOrientation/basic-operation-expected.txt

Tests the basic operation of DeviceOrientation using the mock.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS deviceOrientationEvent.alpha is mockAlpha
PASS deviceOrientationEvent.beta is mockBeta
PASS deviceOrientationEvent.gamma is mockGamma
PASS successfullyParsed is true

TEST COMPLETE
